Services/gRPC/zeta36-chess-alpha-zero/requirements.txt
grpcio>=1.14.2
grpcio-tools>=1.14.1
tensorflow-gpu==1.4
keras==2.1.4
profilehooks
numpy
pyperclip
python-chess
h5py